# apps/bubu-log/openapi.yaml
openapi: 3.1.0
info:
  title: Bubu Log API
  description: Baby activity tracking API
  version: 1.0.0
servers:
  - url: /api
    description: API server
paths:
...
  /voice-input:
    post:
      operationId: parseVoiceInput
      summary: Parse voice input and create activity
      description: |
        Accepts natural language text (from voice input) and uses AI to parse it into a structured activity record.
        Supports Chinese input like "宝宝喝了80毫升奶" or "刚换了尿布，有便便".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: The voice input text to parse
                  example: "宝宝刚才喝了60毫升奶"
              required:
                - text
      responses:
        '201':
          description: Activity created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  activity:
                    $ref: '#/components/schemas/Activity'
                  parsed:
                    type: object
                    properties:
                      confidence:
                        type: number
                        description: AI confidence level (0-1)
                      originalText:
                        type: string
                  message:
                    type: string
                    description: Human-readable confirmation message
        '400':
          description: Parse failed or invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
                    enum: [MISSING_TEXT, PARSE_FAILED, INVALID_TYPE]
                  originalText:
                    type: string
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    ActivityType:
      type: string
      enum:
        - SLEEP
        - DIAPER
        - BREASTFEED
        - BOTTLE
        - PUMP
        - HEAD_LIFT
        - PASSIVE_EXERCISE
        - ROLL_OVER
        - PULL_TO_SIT
        - GAS_EXERCISE
        - BATH
        - OUTDOOR
        - EARLY_EDUCATION
        - SUPPLEMENT
        - SPIT_UP
    MilkSource:
      type: string
      enum:
        - BREAST_MILK
        - FORMULA
    PoopColor:
      type: string
      enum:
        - YELLOW
        - GREEN
        - BROWN
        - BLACK
        - WHITE
        - RED
    PeeAmount:
      type: string
      enum:
        - SMALL
        - MEDIUM
        - LARGE
    BreastFirmness:
      type: string
      enum:
        - SOFT    # 软
        - ELASTIC # 弹
        - HARD    # 硬
    SupplementType:
      type: string
      enum:
        - AD    # 维生素AD
        - D3    # 维生素D3
    SpitUpType:
      type: string
      enum:
        - NORMAL     # 普通吐奶
        - PROJECTILE # 喷射性吐奶
...
    Activity:
      type: object
      required:
        - id
        - type
        - startTime
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/ActivityType'
        startTime:
          type: string
          format: date-time
          description: 活动开始时间
        endTime:
          type: string
          format: date-time
          nullable: true
          description: 活动结束时间（正在进行的活动没有结束时间）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hasPoop:
          type: boolean
          nullable: true
        hasPee:
          type: boolean
          nullable: true
        poopColor:
          $ref: '#/components/schemas/PoopColor'
        poopPhotoUrl:
          type: string
          nullable: true
        peeAmount:
          $ref: '#/components/schemas/PeeAmount'
        burpSuccess:
          type: boolean
          nullable: true
        milkAmount:
          type: integer
          nullable: true
        milkSource:
          $ref: '#/components/schemas/MilkSource'
          nullable: true
        breastFirmness:
          $ref: '#/components/schemas/BreastFirmness'
          nullable: true
        supplementType:
          $ref: '#/components/schemas/SupplementType'
          nullable: true
        spitUpType:
          $ref: '#/components/schemas/SpitUpType'
          nullable: true
        count:
          type: integer
          nullable: true
          description: 次数（用于翻身、拉坐等计数类活动）
        notes:
          type: string
          nullable: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
  responses:
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'